requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
slowapi==0.1.9
pydantic==2.5.0

# Document processing (text-focused)
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
# python-pptx removed - use PresentationLens service instead

# Text analysis
textstat==0.7.3
nltk==3.8.1
scikit-learn==1.4.0

# External services
requests==2.31.0
httpx==0.26.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4